pyproject.toml
[project]
name = "intermediate-wholesaler-credit-creation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymysql>=1.1.2",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
]
